Replace deprecated `setVisibleOps` call

+++ api/net/src/main/kotlin/org/rsmod/api/net/rsprot/RspNpcInfo.kt
@@ -60,7 +60,7 @@ class RspNpcInfo(val rspAvatar: NpcAvatar) : NpcInfoProtocol {
     }
 
     override fun toggleOps(ops: OpVisibility) {
-        rspAvatar.extendedInfo.setVisibleOps(ops.packed)
+        rspAvatar.extendedInfo.setVisibleOps(ops.packed.toByte())
     }
 
     override fun setHeadIcon(slot: Int, graphic: Int, index: Int) {
